Make LICENSE install conditional

+++ PKGBUILD
@@ -50,7 +50,9 @@ package() {
   cd "$pkgname-$pkgver"
   python -m installer --destdir="$pkgdir" dist/*.whl
 
-  install -Dm644 LICENSE "$pkgdir/usr/share/licenses/$pkgname/LICENSE"
+  if [ -f LICENSE ]; then
+    install -Dm644 LICENSE "$pkgdir/usr/share/licenses/$pkgname/LICENSE"
+  fi
 
   if [ -f completions/aegis.bash ]; then
     install -Dm644 completions/aegis.bash \
